% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_theme.R
\name{color_theme}
\alias{color_theme}
\title{Selection of theme color, 1 to 5 for now}
\usage{
color_theme(paleta)
}
\arguments{
\item{paleta}{selection of a theme color}
}
\value{
list of colors
}
\description{
Selection of theme color, 1 to 5 for now
}
\examples{
color_theme(3) # 1 to 7 themes so far
}
