% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_teste_hip_qui_quadrado.R
\name{plot_chi_test}
\alias{plot_chi_test}
\title{Plot Chi-Square test for one populational variance}
\usage{
plot_chi_test(alpha = 0.05, alternative = "two.sided", df, va, vp,
  annotations = TRUE, color = 3)
}
\arguments{
\item{alpha}{significance level for the critical limits}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{df}{degrees of freedom for the sample (default = 10)}

\item{va}{sample variance (default = 10)}

\item{vp}{populational variance under null hypothesis (default = 10)}

\item{annotations}{logical, show annotations in plot figure, TRUE ou FALSE (default = TRUE)}

\item{color}{integer, color theme for plot, from 1 to 7 so far (default = 3)}
}
\description{
Plot Chi-Square test for one populational variance
}
\examples{
plot_chi_test(alpha = 0.05,alternative = "two.sided",
              df = 20, va = 20,
              vp = 15,
              annotations = TRUE,
              color = 3)

}
