% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_teste_hip_F.R
\name{plot_F_test}
\alias{plot_F_test}
\title{Plot F Test to Compare Two Variances}
\usage{
plot_F_test(alpha = 0.05, alternative = "two.sided", df1, df2, v1, v2,
  ratio = 1, annotations = TRUE, color = 3)
}
\arguments{
\item{alpha}{significance level for the critical limits}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{df1}{degrees of freedom for the sample 1 (default = 10)}

\item{df2}{degrees of freedom for the sample 2 (default = 10)}

\item{v1}{sample 1 variance (default = 1.5)}

\item{v2}{sample 2 variance (default = 2.5)}

\item{ratio}{the ratio of population variances under the null hypoyhesis (default = 1)}

\item{annotations}{logical, show annotations in plot figure, TRUE ou FALSE (default = TRUE)}

\item{color}{integer, color theme for plot, from 1 to 7 so far (default = 3)}
}
\description{
Plot F Test to Compare Two Variance, the larger sample variance is always placed in the numerator.
}
\examples{
plot_F_test(alpha = 0.05, alternative = "two.sided",
            df1 = 10, df2 = 10,
            v1 = 1.5, v2 = 2.5,
            ratio = 1,
            annotations = TRUE,
            color = 3)

}
