% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_teste_hip_T.R
\name{plot_T_test}
\alias{plot_T_test}
\title{Plot a two sample t-test of sample statistics}
\usage{
plot_T_test(alpha = 0.05, alternative = "two.sided",
  var.equal = "equal", n1, m1, v1, n2, m2, v2, delta0 = 0,
  annotations = TRUE, color = 3)
}
\arguments{
\item{alpha}{significance level for the critical limits; it must lie between zero and one}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{var.equal}{a char variable indicating whether to treat the two variances as being equal or notequal. If "equal" then the pooled variance is used to estimate the variance otherwise ("notequal") the Welch (or Satterthwaite) approximation to the degrees of freedom is used}

\item{n1}{a single number representing the sample size of sample 1}

\item{m1}{a single number representing the sample mean of sample 1}

\item{v1}{a single number representing the sample variance for sample 1}

\item{n2}{a single number representing the sample size of sample 2}

\item{m2}{a single number representing the sample mean of sample 2}

\item{v2}{a single number representing the sample variance for sample 2}

\item{delta0}{a number indicating the true value of the difference in means of the two sample test.}

\item{annotations}{logical, show annotations in plot figure, TRUE ou FALSE (default = TRUE)}

\item{color}{integer, color theme for plot, from 1 to 7 so far (default = 3)}
}
\description{
Performs two sample t-tests for mean 1 and mean 2 (user supplied summary information). Is only applicable for the 2-sample tests (mu1-mu2).
alternative = "greater" is the alternative that sample mean 1 has a larger mean than sample mean 2.
If var.equal is "equal" then the pooled estimate of the variance is used. By default, if var.equal is "notequal" then the each sample variance is used separately and the Welch modification to the degrees of freedom is used.
Note: The result of Welch modification to the degrees of freedom was truncated to integer.
}
\examples{
plot_T_test(alpha = 0.05, alternative = "two.sided",
            var.equal = "equal",
            n1 = 10, m1 = 3.1, v1 = 1.5,
            n2 = 20, m2 = 4.2, v2 = 1.7,
            delta0 = 0,
            annotations = TRUE,
            color = 3)
}
