% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_plotUD.R
\name{plotBIN}
\alias{plotBIN}
\title{Plot Binomial Discrete Probability Distribution}
\usage{
plotBIN(n, p, hlt = FALSE, i = 2, f = 4)
}
\arguments{
\item{n}{number of trials (zero or more).}

\item{p}{probability of success on each trial.}

\item{hlt}{logical (default is FALSE); if TRUE, probabilities P[i ≤ X ≤ f] are highlighted in green color in plot}

\item{i, f}{highlight probability from i' initial value of distribution to 'f' final value of distribution}
}
\description{
X represent the number of successes in a sequence of `n` Bernoulli trials.
}
\examples{
plotBIN(n = 50, p = 0.15, hlt = TRUE, i = 2, f = 4)
}
