% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_plotUD.R
\name{plotBINNEG}
\alias{plotBINNEG}
\title{Plot Negative Binomial Distribution}
\usage{
plotBINNEG(p, r, n.plot = 50, hlt = FALSE, i = 2, f = 4)
}
\arguments{
\item{p}{}

\item{r}{target for number of successful trials. Must be strictly positive, need not be integer.}

\item{n.plot}{number of observations for the plot, by default will be ploted 50 observations}

\item{hlt}{logical (default is FALSE); if TRUE, probabilities P[i ≤ X ≤ f] are highlighted in green color in the plot}

\item{i, f}{highlight probability from i' initial value of distribution to 'f' final value of distribution}
}
\description{
X represent the number of failures in a sequence of Bernoulli trials before a `r` success occurs.
}
\examples{
plotBINNEG(p = 0.15, r = 3, n.plot = 50 , hlt = TRUE, i = 2, f = 4)
}
