% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_plotUD.R
\name{plotcUD}
\alias{plotcUD}
\title{Plot Cumultaive Uniform Discrete Probability Distribution}
\usage{
plotcUD(a = 0, b = 4)
}
\arguments{
\item{a}{lower limits of the distribution}

\item{b}{upper limits of the distribution}
}
\description{
Plot Cumultaive Uniform Discrete Probability Distribution
}
\examples{
plotcUD(a = 0, b = 4)
}
