% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_plotUD.R
\name{plotGEO}
\alias{plotGEO}
\title{Plot Geometric Distribution}
\usage{
plotGEO(p, n.plot = 50, hlt = FALSE, i = 2, f = 4)
}
\arguments{
\item{p}{probability of success in each trial. 0 < p <= 1.}

\item{n.plot}{number of observations for the plot, by default will be ploted 50 observations}

\item{hlt}{logical (default is FALSE); if TRUE, probabilities P[i ≤ X ≤ f] are highlighted in green color in the plot}

\item{i, f}{highlight probability from i' initial value of distribution to 'f' final value of distribution}
}
\description{
X represent the number of failures in a sequence of Bernoulli trials before a success occurs.
}
\examples{
plotGEO(p = 0.15, n.plot = 50 , hlt = FALSE, i = 2, f = 4)
}
