% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_plotUD.R
\name{plotPOIS}
\alias{plotPOIS}
\title{Plot Poisson Distribution}
\usage{
plotPOIS(lambda, n.plot = 50, hlt = FALSE, i = 2, f = 4)
}
\arguments{
\item{lambda}{is the mean number of events per unit length}

\item{n.plot}{number of observations for the plot, by default will be ploted 50 observations}

\item{hlt}{logical (default is FALSE); if TRUE, probabilities P[i ≤ X ≤ f] are highlighted in green color in the plot}

\item{i, f}{highlight probability from i' initial value of distribution to 'f' final value of distribution}
}
\description{
Plot Poisson Distribution
}
\examples{
plotPOIS(lambda = 0.5, n.plot = 15, hlt = TRUE, i = 2, f = 4)
}
